Keep only properties in Flight.cs to be Model and relocate its path
remove TicketClass enum, prices dictionary, and Flight constructor
